+++ common/posicion.h
@@ -7,14 +7,14 @@
 #include <string>
 
 typedef struct Posicion {
-    int coordenada_x;
-    int coordenada_y;
+    float coordenada_x;
+    float coordenada_y;
 
     // Constructor por defecto
     Posicion(): coordenada_x(0), coordenada_y(0) {}
 
     // Constructor por parámetros
-    Posicion(int x, int y): coordenada_x(x), coordenada_y(y) {}
+    Posicion(float x, float y): coordenada_x(x), coordenada_y(y) {}
 
     // Constructor de copia
     Posicion(const Posicion& pos): coordenada_x(pos.coordenada_x), coordenada_y(pos.coordenada_y) {}
@@ -28,7 +28,7 @@ typedef struct Posicion {
         return *this;
     }
 
-    void set_posicion(int x, int y) {
+    void set_posicion(float x, float y) {
         coordenada_x = x;
         coordenada_y = y;
     }
